minor: refactoring the ssl test suites to remove some redundancy
